Add empty check of heartbeats base url
When the client is initializing there is an interval of time where the Start method doesn't get called yet but the coroutine SendHeartBeatRequest does though. 

In that interval, the client makes a web request with an empty string as url which gives an "curl error 3 url malformed" error.

This check avoid that situation.