Don't continue heartbeat thread on ExecutionException.
In the context of the heartbeat thread, an ExecutionException means that
something unexpected went wrong when trying to send a heartbeat to the
agent.  In that case, we should force the process to exit because
we can no longer reliably communicate with the Agent.

We should also let the game server know that we're about to shutdown
by invoking the shutdown callback.