Java: Keep heartbeat thread alive (#62)
* Java: Keep heartbeat thread alive.

Previously, if something went wrong on the heartbeat thread we would
simply convert the checked exception into a runtime exception and then
throw it.  That would ultimately terminate our heartbeat thread, but
the game server process would have no idea.  The end result is that
game servers could end up stuck in the `readyForPlayers()` method
forever.

This approach catches only the exceptions that we expect, logs them and
then keeps going.  If we encounter an exception that we can't handle
then we log it and then end the process since we can no longer safely
reason about the state of our application.

Lastly, I shuffled around the location where we build the URI for
the agent endpoint out of the `sendHeartbeat` method so that we didn't
have to deal with the `URISyntaxException` in our main loop.  I figure
that since that URI doesn't change with time, then we should be able
to just build it once.

* Don't continue heartbeat thread on ExecutionException.

In the context of the heartbeat thread, an ExecutionException means that
something unexpected went wrong when trying to send a heartbeat to the
agent.  In that case, we should force the process to exit because
we can no longer reliably communicate with the Agent.

We should also let the game server know that we're about to shutdown
by invoking the shutdown callback.