Dynamically adjust the heartbeat interval according to VMAgent configuration
Prior to this change, the GSDK would send a heartbeat to VMAgent at a fixed rate once per second. With this change, it now parses the nextHeartbeatIntervalMs field from the heartbeat response to dynamically adjust the rate at which it heartbeats. This can help reduce VMAgent CPU usage in game configurations with high numbers of servers per VM (e.g. > 200 servers per VM).