Fix ConfigMutex to cover all accesses to ConfigSettings
Remove direct access to ConfigMutex and ConfigSettings, and instead use a single key accessor which applies the lock properly.
Reconfigure all accesses to utilize the single key accessor (matches the use-case in all locations)
Also a minor improvement in the DecodeHeartbeatResponse to shorten/eliminate the time that ConfigMutex lock is held.